requirements.txt
numpy
git+https://github.com/GlobalFishingWatch/bqtools.git
pytz
python-dateutil
